+++ tap_square/__init__.py
@@ -3,6 +3,7 @@ import json
 
 import singer
 from singer.catalog import write_catalog
+from tap_square.client import SquareClient
 from tap_square.discover import discover
 from tap_square.sync import sync
 
@@ -15,12 +16,15 @@ def main():
     is_sandbox = args.config.get('sandbox')
     if isinstance(is_sandbox, str):
         is_sandbox = args.config.get('sandbox') == 'true'
-    catalog = args.catalog if args.catalog else discover(is_sandbox)
+
+    client = SquareClient(args.config, args.config_path)
 
     if args.discover:
+        catalog = discover(client, is_sandbox)
         write_catalog(catalog)
     else:
-        sync(args.config, args.config_path, args.state, catalog)
+        catalog = args.catalog if args.catalog else discover(client, is_sandbox)
+        sync(args.config, args.config_path, args.state, catalog, client)
 
 if __name__ == '__main__':
     main()