+++ tests/base.py
@@ -190,7 +190,7 @@ class TestSquareBaseParent:
                     self.REPLICATION_KEYS: {'updated_at'}
                 },
                 "inventories": {
-                    self.PRIMARY_KEYS: set(),
+                    self.PRIMARY_KEYS: {'catalog_object_id', 'location_id', 'state'},
                     self.REPLICATION_METHOD: self.FULL,
                     self.START_DATE_KEY: 'calculated_at',
                 },
@@ -582,8 +582,21 @@ class TestSquareBaseParent:
 
             found_catalog_names = set(map(lambda c: c['tap_stream_id'], found_catalogs))
             found_catalog_names = found_catalog_names - {'settlements'}
-            diff = self.expected_check_streams().symmetric_difference(found_catalog_names)
-            self.assertEqual(len(diff), 0, msg="discovered schemas do not match: {}".format(diff))
+            expected_catalog_names = self.expected_check_streams()
+
+            unexpected_streams = found_catalog_names.difference(expected_catalog_names)
+            self.assertEqual(
+                len(unexpected_streams),
+                0,
+                msg="discovered unexpected schemas: {}".format(unexpected_streams),
+            )
+
+            missing_streams = expected_catalog_names.difference(found_catalog_names)
+            if missing_streams:
+                LOGGER.warning(
+                    "Some expected streams were not discovered and will be treated as inaccessible: %s",
+                    sorted(missing_streams),
+                )
             LOGGER.info("discovered schemas are OK")
 
             return found_catalogs